+++ tests/integration/cli.edge-cases.test.js
@@ -162,64 +162,4 @@ describe('integration.cli.edge-cases', () => {
       `${outFile} should equal ${expectedOutSnap}`
     )
   }, cliRunTestTimeout)
-
-  describe('npm-version depending npm-args', () => {
-    const tmpRootRun = join(tmpRoot, 'npmVersion-depending-npmArgs')
-    mkdirSync(tmpRootRun)
-
-    const rMinor = Math.round(99 * Math.random())
-    const rPatch = Math.round(99 * Math.random())
-    const le6 = Math.round(6 * Math.random())
-    const ge7 = 7 + Math.round(92 * Math.random())
-
-    const npmArgsGeneral = ['--json', '--long']
-    const npm6ArgsGeneral = [...npmArgsGeneral, '--depth=255']
-    const npm7ArgsGeneral = [...npmArgsGeneral, '--all']
-    const npm8ArgsGeneral = [...npmArgsGeneral, '--all']
-    const npm9ArgsGeneral = [...npmArgsGeneral, '--all']
-    const npm10ArgsGeneral = [...npmArgsGeneral, '--all']
-
-    test.each([
-      ['basic npm 6', `6.${rMinor}.${rPatch}`, [], npm6ArgsGeneral],
-      ['basic npm 7', `7.${rMinor}.${rPatch}`, [], npm7ArgsGeneral],
-      ['basic npm 8', `8.${rMinor}.${rPatch}`, [], npm8ArgsGeneral],
-      ['basic npm 9', `9.${rMinor}.${rPatch}`, [], npm9ArgsGeneral],
-      ['basic npm 10', `10.${rMinor}.${rPatch}`, [], npm10ArgsGeneral],
-      // region omit
-      ['omit everything npm 6', `6.${rMinor}.${rPatch}`, ['--omit', 'dev', 'optional', 'peer'], [...npm6ArgsGeneral, '--production']],
-      ['omit everything npm 7', `7.${rMinor}.${rPatch}`, ['--omit', 'dev', 'optional', 'peer'], [...npm7ArgsGeneral, '--production']],
-      ['omit everything npm lower 8.7', `8.${le6}.${rPatch}`, ['--omit', 'dev', 'optional', 'peer'], [...npm8ArgsGeneral, '--production']],
-      ['omit everything npm greater-equal 8.7', `8.${ge7}.${rPatch}`, ['--omit', 'dev', 'optional', 'peer'], [...npm8ArgsGeneral, '--omit=dev', '--omit=optional', '--omit=peer']],
-      ['omit everything npm 9', `9.${rMinor}.${rPatch}`, ['--omit', 'dev', 'optional', 'peer'], [...npm9ArgsGeneral, '--omit=dev', '--omit=optional', '--omit=peer']],
-      ['omit everything npm 10', `10.${rMinor}.${rPatch}`, ['--omit', 'dev', 'optional', 'peer'], [...npm10ArgsGeneral, '--omit=dev', '--omit=optional', '--omit=peer']],
-      // endregion
-      // region package-lock-only
-      ['package-lock-only not supported npm 6 ', `6.${rMinor}.${rPatch}`, ['--package-lock-only'], [...npm6ArgsGeneral]],
-      ['package-lock-only npm 7', `7.${rMinor}.${rPatch}`, ['--package-lock-only'], [...npm7ArgsGeneral, '--package-lock-only']],
-      ['package-lock-only npm 8', `8.${rMinor}.${rPatch}`, ['--package-lock-only'], [...npm8ArgsGeneral, '--package-lock-only']],
-      ['package-lock-only npm 9', `9.${rMinor}.${rPatch}`, ['--package-lock-only'], [...npm9ArgsGeneral, '--package-lock-only']],
-      ['package-lock-only npm 10', `10.${rMinor}.${rPatch}`, ['--package-lock-only'], [...npm10ArgsGeneral, '--package-lock-only']]
-      // endregion
-    ])('%s', async (purpose, npmVersion, cdxArgs, expectedArgs) => {
-      const logFileBase = join(tmpRootRun, purpose.replace(/\W/g, '_'))
-      const cwd = dummyProjectsRoot
-
-      const { res, errFile } = runCLI([
-        ...cdxArgs,
-        '--',
-        join('with-lockfile', 'package.json')
-      ], logFileBase, cwd, {
-        CT_VERSION: npmVersion,
-        CT_EXPECTED_ARGS: expectedArgs.join(' '),
-        npm_execpath: npmLsReplacement.checkArgs
-      })
-
-      try {
-        await expect(res).resolves.toBe(0)
-      } catch (err) {
-        process.stderr.write(readFileSync(errFile))
-        throw err
-      }
-    }, cliRunTestTimeout)
-  })
 })
